update project feature has been added.

+++ src/app/modules/project/project.controller.ts
@@ -4,6 +4,7 @@ import { TProject } from "./project.interface";
 import ProjectServices from "./project.service";
 import { sendResponse } from "../../utils/sendResponse";
 import statusCode from "../../utils/statusCodes";
+import AppError from "../../utils/AppError";
 
 const createProject = catchAsync(async (req: Request, res: Response) => {
   const payload: { data: TProject; file: Express.Multer.File } = {
@@ -34,6 +35,9 @@ const getAllProjects = catchAsync(async (req: Request, res: Response) => {
 
 const getProjectById = catchAsync(async (req: Request, res: Response) => {
   const { id } = req.params;
+  if (!id) {
+    throw new AppError(statusCode.NOT_FOUND, "Please provide an id.");
+  }
   const result = await ProjectServices.getProjectById(id);
 
   sendResponse(res, {
@@ -44,10 +48,11 @@ const getProjectById = catchAsync(async (req: Request, res: Response) => {
   });
 });
 
-
 const deleteByID = catchAsync(async (req: Request, res: Response) => {
-
   const { id } = req.params;
+  if (!id) {
+    throw new AppError(statusCode.NOT_FOUND, "Please provide an id.");
+  }
   const result = await ProjectServices.deleteByID(id);
 
   sendResponse(res, {
@@ -56,7 +61,25 @@ const deleteByID = catchAsync(async (req: Request, res: Response) => {
     message: "Project Deleted successfully",
     data: result,
   });
+});
 
+const update = catchAsync(async (req: Request, res: Response) => {
+
+  const { id } = req.params;
+  if (!id) {
+    throw new AppError(statusCode.NOT_FOUND, "Please provide an id.");
+  }
+
+  const payload: { data: TProject; file: Express.Multer.File } = {data: req.body, file: req.file!};
+
+  const result = await ProjectServices.update(id, payload);
+
+  sendResponse(res, {
+    statusCode: statusCode.OK,
+    success: true,
+    message: "Project Updated successfully",
+    data: result,
+  });
 });
 
 const ProjectController = {
@@ -64,6 +87,7 @@ const ProjectController = {
   getAllProjects,
   getProjectById,
   deleteByID,
+  update,
 };
 
 export default ProjectController;
